requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
opencv-python
onnxruntime
python-dotenv
pydantic-settings